# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carlos BuckHill Api",
        "description": "L5 Swagger Api documentation for Buckhill Challenge",
        "contact": {
            "email": "quiqueal.19@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8085/api/v1/",
            "description": "Demo API/V1 Server for challenge"
        }
    ],
    "paths": {
...
        "/order": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all personal orders",
                "operationId": "2baa0fb4d65d150b378f590e0665d91c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new Order",
                "operationId": "bcd22b48f5dbef32b92f050cfdd9590a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_status_id": {
                                        "type": "uuid"
                                    },
                                    "products": {
                                        "type": "object",
                                        "example": "[{'product': 'product_uuid', 'quantity': '1'}]"
                                    },
                                    "delivery_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "22.5"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully"
                    },
                    "201": {
                        "description": "Order created successfully"
                    },
                    "404": {
                        "description": "Not created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "operationId": "a69b9fa27fd7579b888b5c864937933c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Bearer token authorization",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}